IOS-4815 Use wrapper view for publishing
