+++ Anytype/Sources/PresentationLayer/Settings/PublishToWeb/PublishToWebViewModel.swift
@@ -3,30 +3,22 @@ import Combine
 import SwiftUI
 import Services
 import AnytypeCore
+import ProtobufMessages
 
-enum PublishToWebViewModelState {
+enum PublishToWebViewModelState: Equatable {
+    case initial
     case error(String)
-    case ok
-    
-    var isError: Bool {
-       if case .error = self { return true }
-       return false
-   }
+    case loaded(PublishToWebViewInternalData)
 }
 
 @MainActor
 final class PublishToWebViewModel: ObservableObject {
-    
-    @Published var customPath: String = ""
-    @Published var showJoinSpaceButton: Bool = true
-    @Published var canPublish: Bool = true
-    
-    @Published var state = PublishToWebViewModelState.ok
-    
-    let domain: String
+    @Published var state = PublishToWebViewModelState.initial
     
     @Injected(\.publishingService)
     private var publishingService: any PublishingServiceProtocol
+    @Injected(\.accountParticipantsStorage)
+    private var participantStorage: any AccountParticipantsStorageProtocol
     
     private let spaceId: String
     private let objectId: String
@@ -34,31 +26,31 @@ final class PublishToWebViewModel: ObservableObject {
     init(data: PublishToWebViewData) {
         spaceId = data.spaceId
         objectId = data.objectId
-        
-        let participantStorage = Container.shared.accountParticipantsStorage()
-        if let participant = participantStorage.participants.first {
-            domain = participant.publishingDomain
-        } else {
-            anytypeAssertionFailure("No participants found for account")
-            domain = ""
-            state = .error(Loc.Publishing.Error.noDomain)
-        }
-        
-        setupBindings()
     }
     
     func onAppear() async {
-//        let status = try? await publishingService.getStatus(spaceId: spaceId, objectId: objectId)
-        // TBD;
-    }
-    
-    func onPublishTap() {
-        // TBD;
-    }
-    
-    private func setupBindings() {
-        $customPath
-            .map { !$0.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty }
-            .assign(to: &$canPublish)
+        guard let domain = participantStorage.participants.first?.publishingDomain else {
+            anytypeAssertionFailure("No participants found for account")
+            state = .error(Loc.Publishing.Error.noDomain)
+            return
+        }
+        
+        do {
+            let status: PublishState?
+            if let newStatus = try await publishingService.getStatus(spaceId: spaceId, objectId: objectId) {
+                status = newStatus
+            } else {
+                status = nil
+            }
+            
+            state = .loaded(PublishToWebViewInternalData(
+                objectId: objectId,
+                spaceId: spaceId,
+                domain: domain,
+                status: status
+            ))
+        } catch {
+            state = .error(error.localizedDescription)
+        }
     }
 }
